feat: order list in dashboard lab
