+++ src/router/routes/lab.js
@@ -11,7 +11,11 @@ const labRoutes = [
         path: "/",
         name: "lab-dashboard",
         meta: {
-          pageHeader: "Lab Dashboard"
+          pageHeader: "Lab Dashboard",
+          breadcrumbs: [
+            { text: "Lab", href: "/lab" },
+            { text: "Orders", disabled: true }
+          ]
         },
         component: () => import(/* webpackChunkName */ "@/views/Dashboard/Lab"),
         beforeEnter: (to, from, next) => {
