+++ src/views/Dashboard/Lab/index.vue
@@ -13,6 +13,9 @@
 .lab-verified-box {
   display: flex;
   align-items: center;
+  background: #fff;
+  padding: 20px;
+  border-radius: 5px;
 }
 .lab-verified-box > * {
   display: block;
@@ -31,71 +34,37 @@
         {{ computeStakingStatus ? "Continue Registration" : "Register Now!" }}
       </v-btn>
     </v-container>
-    <v-row v-else justify="center" class="pr20">
-      <v-col cols="12" md="10">
-        <v-card class="dg-card" elevation="0" outlined>
-          <div class="secondary--text mt-5 ml-8 mb-5 lab-verified-box">
-            <v-icon 
-              inline
-              color="primary"
-              :size="20"
-              v-if="labAccount && labAccount.verificationStatus == 'Verified'"
-            >mdi-check-decagram</v-icon>
-            <v-icon 
-              inline
-              color="yellow darken-2"
-              :size="20"
-              v-else
-            >mdi-information</v-icon>
-            <b v-if="labAccount && labAccount.verificationStatus == 'Unverified'">Your verification submission is being reviewed by DAOGenics</b>
-            <b v-else>{{ computeVerificationStatus }}</b>
-          </div>
-        </v-card>
-      </v-col>
-      <v-col cols="12" md="5">
-        <MenuCard
-          :height="120"
-          class="mb-2"
-          icon="mdi-microscope"
-          title="Account"
-          sub-title="Account page"
-          :to="{ name: 'lab-dashboard-account' }"
-        />
-        <MenuCard
-          :height="120"
-          class="mb-2"
-          icon="mdi-hand-heart"
-          title="Services"
-          sub-title="Services list"
-          :to="{ name: 'lab-dashboard-services' }"
-        />
-        <MenuCard
-          :height="120"
-          class="mb-2"
-          icon="mdi-clipboard-plus"
-          title="Orders"
-          sub-title="Payment History"
-          :to="{ name: 'lab-dashboard-order-history' }"
-        />
-      </v-col>
-      <v-col cols="12" md="5">
-        <LabOrders />
-      </v-col>
-    </v-row>
+    <div v-else>
+      <div class="secondary--tex mb-5 lab-verified-box" v-if="labAccount.verificationStatus !== 'Verified'">
+        <v-icon 
+          inline
+          color="primary"
+          :size="20"
+          v-if="labAccount && labAccount.verificationStatus == 'Verified'"
+        >mdi-check-decagram</v-icon>
+        <v-icon 
+          inline
+          color="yellow darken-2"
+          :size="20"
+          v-else
+        >mdi-information</v-icon>
+        <b v-if="labAccount && labAccount.verificationStatus == 'Unverified'">Your verification submission is being reviewed by DAOGenics</b>
+        <b v-else>{{ computeVerificationStatus }}</b>
+      </div>
+      <OrderList isDashboard v-else />
+    </div>
   </v-container>
 </template>
 
 <script>
 import { mapState } from "vuex"
-import MenuCard from "@/components/MenuCard"
-import LabOrders from "./LabOrders";
+import OrderList from "@/components/OrderList";
 
 export default {
   name: "Lab",
 
   components: {
-    MenuCard,
-    LabOrders
+    OrderList
   },
 
   computed: {
